requirements.txt
hydra-core
hydra_colorlog

tornado
tornado-swagger

redisearch
# dont delete this line